requirements.txt
click==8.0.1
numpy==1.19.1
pandas==1.1.0
torch==1.9.0
torchvision==0.10.0
torchaudio==0.9.0
catalyst==21.4.2
python-Levenshtein==0.12.2
tqdm==4.60.0
scikit-learn==0.23.2
audiomentations==0.17.0